pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "openai>=1.99.3",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
